Start of List and Task implementation
