+++ src/main/java/org/example/todo_list/SceneManager.java
@@ -9,7 +9,7 @@ import java.io.IOException;
 import java.util.HashMap;
 
 public class SceneManager {
-    private static final SceneManager INSTANCE  = new SceneManager();;
+    private static final SceneManager INSTANCE  = new SceneManager();
     /**
      * Holds loaded scenes that can be shown.
      */
@@ -46,8 +46,8 @@ public class SceneManager {
             scenes.put(fxmlSceneName, scene);
             return true;
         } catch (IOException e) {
-            //e.printStackTrace();
-            System.out.println("Failed to load scene "+fxmlSceneName);
+            e.printStackTrace();
+            System.out.println("Failed to load scene "+fxmlSceneName+".fxml");
             return false;
         }
     }
