+++ src/main/java/org/example/todo_list/view_models/ListController.java
@@ -1,0 +1,7 @@
+package org.example.todo_list.view_models;
+
+public class ListController {
+
+    //Todo make List name Editable.
+
+}
