+++ src/main/java/org/example/todo_list/view_models/LoginController.java
@@ -51,7 +51,6 @@ public class LoginController implements Initializable {
 
     public void signIn(ActionEvent actionEvent) {
         //Todo check and validate user info from Database.
-        sceneManager.loadScene("HomeScene");
         sceneManager.showScene("HomeScene");
     }
