+++ src/main/java/org/example/todo_list/view_models/TaskController.java
@@ -1,0 +1,7 @@
+package org.example.todo_list.view_models;
+
+public class TaskController {
+
+    //Todo make Task name's text clickable to check off Task.
+
+}
