+++ src/main/resources/org/example/todo_list/css/taskStyle.css
@@ -1,0 +1,15 @@
+.text {
+    -fx-font-family: Arial;
+}
+
+.toggle-button {
+    -fx-background-image: url("./../images/menu_icons/ellipse-outline_32px.png");
+    -fx-background-repeat: no-repeat;
+    -fx-background-size: 20px;
+    -fx-max-width: 20px;
+    -fx-max-height: 20px;
+    -fx-background-color: transparent;
+}
+.toggle-button:selected {
+    -fx-background-image: url('./../images/menu_icons/checkmark-circle-outline_32px.png');
+}
