+++ src/main/resources/org/example/todo_list/images/logo/icon_32px.png
